add init_sod
