+++ tests/sod.rs
@@ -3,7 +3,7 @@
 // License: MIT
 
 use color_eyre::{eyre::Context, Result};
-use corries::prelude::*;
+use corries::{initfuncs::init_sod, prelude::*};
 const S: usize = 100;
 
 fn get_config<N: NumFlux<E, S> + 'static, const E: usize>(folder_name: &str, file_name: &str) -> CorriesConfig {
@@ -25,29 +25,6 @@ fn get_config<N: NumFlux<E, S> + 'static, const E: usize>(folder_name: &str, fil
     };
 }
 
-fn init<P, N, T, const E: usize, const S: usize>(
-    u: &mut State<P, E, S>,
-    _: &mut Solver<P, N, T, E, S>,
-    _: &Mesh<S>,
-) -> Result<()>
-where
-    P: Physics<E, S>,
-    N: NumFlux<E, S>,
-    T: TimeSolver<P, E, S>,
-{
-    let breakpoint_index = (S as f64 * 0.5) as usize;
-    u.cent.prim.fill(0.0);
-    for i in 0..breakpoint_index {
-        u.cent.prim[[P::JRHO, i]] = 1.0;
-        u.cent.prim[[P::JPRESSURE, i]] = 1.0;
-    }
-    for i in breakpoint_index..S {
-        u.cent.prim[[P::JRHO, i]] = 0.125;
-        u.cent.prim[[P::JPRESSURE, i]] = 0.1;
-    }
-    Ok(())
-}
-
 #[test]
 fn sod_hll() -> Result<()> {
     set_Physics_and_E!(Euler1DAdiabatic);
@@ -55,7 +32,7 @@ fn sod_hll() -> Result<()> {
     type T = RungeKuttaFehlberg<P, E, S>;
 
     get_config::<N, E>("results/integrationtests/sod_hll", "sod_hll")
-        .init_corries::<P, N, T, E, S>(init)
+        .init_corries::<P, N, T, E, S>(init_sod)
         .context("While calling CorriesConfig::init_corries")?
         .run_corries()
 }
@@ -67,7 +44,7 @@ fn sod_kt() -> Result<()> {
     type T = RungeKuttaFehlberg<P, E, S>;
 
     get_config::<N, E>("results/integrationtests/sod_kt", "sod_kt")
-        .init_corries::<P, N, T, E, S>(init)
+        .init_corries::<P, N, T, E, S>(init_sod)
         .context("While calling CorriesConfig::init_corries")?
         .run_corries()
 }
